src/main/java/com/rtg/variant/cnv/segment/EnergySegmentScorer.java
/*
 * Copyright (c) 2018. Real Time Genomics Limited.
 *
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are
 * met:
 *
 * 1. Redistributions of source code must retain the above copyright
 *    notice, this list of conditions and the following disclaimer.
 *
 * 2. Redistributions in binary form must reproduce the above copyright
 *    notice, this list of conditions and the following disclaimer in the
 *    documentation and/or other materials provided with the
 *    distribution.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
 * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
 * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
 * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
package com.rtg.variant.cnv.segment;

/**
 * Energy based scorer.
 */
class EnergySegmentScorer implements SegmentScorer {
  private final double mAlpha;
  private final double mAleph;

  EnergySegmentScorer(final double alpha, final double aleph) {
    mAlpha = alpha;
    mAleph = aleph;
  }

  @Override
  public double score(final Segment a, final Segment b) {
    final double m = a.mean() - b.mean();
    final double al = a.bins();
    final double bl = b.bins();
    // Original algorithm has log(|d_1-d_2|), 1+ to avoid potential singularity
    return (al * bl / (al + bl)) * m * m
      + mAlpha * Math.log(1 + Math.abs(a.meanDistanceBetween() - b.meanDistanceBetween()))
      + mAleph * Math.log(1 + b.distanceToPrevious());
  }
}
